re order migrations

+++ db/migrate/20240906082659_dropuser.rb
@@ -1,5 +1,7 @@
 class Dropuser < ActiveRecord::Migration[7.1]
   def change
-    drop_table :users
+    if table_exists?(:users)
+      drop_table :users
+    end
   end
 end
